requirements.txt
Django>=5.0
djangorestframework
requests
markdown
django-filter
